+++ projects/storefrontlib/cms-structure/lcp-context/index.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './cms-lcp.service';
+export * from '../services/cms-lcp.service';
 export * from './lcp-context-directive.module';
 export * from './lcp-context.directive';
 export * from './lcp-context.model';
